backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
google-generativeai
youtube-transcript-api
PyPDF2
langchain-text-splitters
python-dotenv
